+++ cli/src/pcluster/config/cluster_config.py
@@ -169,7 +169,7 @@ from pcluster.validators.instances_validators import (
     InstancesAllocationStrategyValidator,
     InstancesCPUValidator,
     InstancesEFAValidator,
-    InstancesMemorySchedulingValidator,
+    InstancesMemorySchedulingWarningValidator,
     InstancesNetworkingValidator,
 )
 from pcluster.validators.kms_validators import KmsKeyIdEncryptedValidator, KmsKeyValidator
@@ -203,6 +203,7 @@ from pcluster.validators.slurm_settings_validator import (
     CustomSlurmSettingLevel,
     CustomSlurmSettingsIncludeFileOnlyValidator,
     CustomSlurmSettingsValidator,
+    SlurmNodePrioritiesWarningValidator,
 )
 from pcluster.validators.tags_validators import ComputeResourceTagsValidator
 
@@ -759,16 +760,23 @@ class SchedulerPluginQueueNetworking(SlurmQueueNetworking):
     pass
 
 
-class Ssh(Resource):
-    """Represent the SSH configuration for a node."""
+class _BaseSsh(Resource):
+    """Represent the base SSH configuration, with the fields in common between all the Ssh."""
 
-    def __init__(self, key_name: str = None, allowed_ips: str = None, **kwargs):
+    def __init__(self, key_name: str = None, **kwargs):
         super().__init__(**kwargs)
         self.key_name = Resource.init_param(key_name)
-        self.allowed_ips = Resource.init_param(allowed_ips, default=CIDR_ALL_IPS)
 
     def _register_validators(self, context: ValidatorContext = None):  # noqa: D102 #pylint: disable=unused-argument
         self._register_validator(KeyPairValidator, key_name=self.key_name)
+
+
+class HeadNodeSsh(_BaseSsh):
+    """Represent the SSH configuration for HeadNode."""
+
+    def __init__(self, allowed_ips: str = None, **kwargs):
+        super().__init__(**kwargs)
+        self.allowed_ips = Resource.init_param(allowed_ips, default=CIDR_ALL_IPS)
 
 
 class Dcv(Resource):
@@ -911,19 +919,17 @@ class S3Access(Resource):
             return [self.bucket_name, f"{self.bucket_name}/*"]
 
 
-class Iam(Resource):
-    """Represent the IAM configuration for HeadNode and Queue."""
+class _BaseIam(Resource):
+    """Represent the base IAM configuration, with the fields in common between all the Iams."""
 
     def __init__(
         self,
-        s3_access: List[S3Access] = None,
         additional_iam_policies: List[AdditionalIamPolicy] = (),
         instance_role: str = None,
         instance_profile: str = None,
         **kwargs,
     ):
         super().__init__(**kwargs)
-        self.s3_access = s3_access
         self.additional_iam_policies = additional_iam_policies
         self.instance_role = Resource.init_param(instance_role)
         self.instance_profile = Resource.init_param(instance_profile)
@@ -968,6 +974,28 @@ class Iam(Resource):
             self._register_validator(RoleValidator, role_arn=self.instance_role)
         elif self.instance_profile:
             self._register_validator(InstanceProfileValidator, instance_profile_arn=self.instance_profile)
+
+
+class Iam(_BaseIam):
+    """Represent the IAM configuration for HeadNode and Queue."""
+
+    def __init__(
+        self,
+        s3_access: List[S3Access] = None,
+        **kwargs,
+    ):
+        super().__init__(**kwargs)
+        self.s3_access = s3_access
+
+
+class LoginNodesIam(_BaseIam):
+    """Represent the IAM configuration for LoginNodes."""
+
+    def __init__(
+        self,
+        **kwargs,
+    ):
+        super().__init__(**kwargs)
 
 
 class Imds(Resource):
@@ -1182,6 +1210,76 @@ class CustomActions(Resource):
         self.on_node_updated = Resource.init_param(on_node_updated)
 
 
+class LoginNodesImage(Resource):
+    """Represent the Image configuration of LoginNodes."""
+
+    def __init__(self, custom_ami: str):
+        super().__init__()
+        self.custom_ami = Resource.init_param(custom_ami)
+
+    def _register_validators(self, context: ValidatorContext = None):  # noqa: D102 #pylint: disable=unused-argument
+        if self.custom_ami:
+            self._register_validator(CustomAmiTagValidator, custom_ami=self.custom_ami)
+
+
+class LoginNodesSsh(_BaseSsh):
+    """Represent the SSH configuration for LoginNodes."""
+
+    def __init__(self, **kwargs):
+        super().__init__(**kwargs)
+
+
+class LoginNodesNetworking(_BaseNetworking):
+    """Represent the networking configuration for LoginNodes."""
+
+    def __init__(
+        self,
+        subnet_id: str,
+        **kwargs,
+    ):
+        super().__init__(**kwargs)
+        self.subnet_id = Resource.init_param(subnet_id)
+
+
+class LoginNodesPools(Resource):
+    """Represent the configuration of a LoginNodePool."""
+
+    def __init__(
+        self,
+        name: str,
+        instance_type: str,
+        image: LoginNodesImage = None,
+        networking: LoginNodesNetworking = None,
+        count: int = None,
+        ssh: LoginNodesSsh = None,
+        iam: LoginNodesIam = None,
+        **kwargs,
+    ):
+        super().__init__(**kwargs)
+        self.name = Resource.init_param(name)
+        self.instance_type = Resource.init_param(instance_type)
+        self.image = image
+        self.networking = networking
+        self.count = Resource.init_param(count, default=1)
+        self.ssh = ssh
+        self.iam = iam or LoginNodesIam(implied=True)
+
+    def _register_validators(self, context: ValidatorContext = None):  # noqa: D102 #pylint: disable=unused-argument
+        self._register_validator(InstanceTypeValidator, instance_type=self.instance_type)
+
+
+class LoginNodes(Resource):
+    """Represent the configuration of LoginNodes."""
+
+    def __init__(
+        self,
+        pools: List[LoginNodesPools],
+        **kwargs,
+    ):
+        super().__init__(**kwargs)
+        self.pools = pools
+
+
 class HeadNode(Resource):
     """Represent the Head Node resource."""
 
@@ -1189,7 +1287,7 @@ class HeadNode(Resource):
         self,
         instance_type: str,
         networking: HeadNodeNetworking,
-        ssh: Ssh = None,
+        ssh: HeadNodeSsh = None,
         disable_simultaneous_multithreading: bool = None,
         local_storage: LocalStorage = None,
         dcv: Dcv = None,
@@ -1204,7 +1302,7 @@ class HeadNode(Resource):
             disable_simultaneous_multithreading, default=False
         )
         self.networking = networking
-        self.ssh = ssh or Ssh(implied=True)
+        self.ssh = ssh or HeadNodeSsh(implied=True)
         self.local_storage = local_storage or LocalStorage(implied=True)
         self.dcv = dcv
         self.custom_actions = custom_actions
@@ -1941,6 +2039,8 @@ class _BaseSlurmComputeResource(BaseComputeResource):
         health_checks: HealthChecks = None,
         custom_slurm_settings: Dict = None,
         tags: List[Tag] = None,
+        static_node_priority: int = None,
+        dynamic_node_priority: int = None,
         **kwargs,
     ):
         super().__init__(**kwargs)
@@ -1959,6 +2059,8 @@ class _BaseSlurmComputeResource(BaseComputeResource):
         self.health_checks = health_checks or HealthChecks(implied=True)
         self.custom_slurm_settings = Resource.init_param(custom_slurm_settings, default={})
         self.tags = tags
+        self.static_node_priority = Resource.init_param(static_node_priority, default=1)
+        self.dynamic_node_priority = Resource.init_param(dynamic_node_priority, default=1000)
 
     @staticmethod
     def fetch_instance_type_info(instance_type) -> InstanceTypeInfo:
@@ -2040,6 +2142,21 @@ class SlurmFlexibleComputeResource(_BaseSlurmComputeResource):
             flexible_instance_type.instance_type
             for flexible_instance_type in self.instances  # pylint: disable=not-an-iterable
         ]
+
+    def _min_schedulable_memory_and_instance_type(self):
+        instances_and_memory = {t: info.ec2memory_size_in_mib() for t, info in self.instance_type_info_map.items()}
+        smallest_type = min(instances_and_memory, key=instances_and_memory.get)
+        return smallest_type, instances_and_memory[smallest_type]
+
+    def _register_validators(self, context: ValidatorContext = None):
+        super()._register_validators(context)
+        smallest_type, min_memory = self._min_schedulable_memory_and_instance_type()
+        self._register_validator(
+            SchedulableMemoryValidator,
+            schedulable_memory=self.schedulable_memory,
+            ec2memory=min_memory,
+            instance_type=smallest_type,
+        )
 
     @property
     def disable_simultaneous_multithreading_manually(self) -> bool:
@@ -2319,6 +2436,11 @@ class SlurmQueue(_CommonQueue):
             queue_name=self.name,
             subnet_ids=self.networking.subnet_ids,
             az_subnet_ids_mapping=self.networking.az_subnet_ids_mapping,
+        )
+        self._register_validator(
+            SlurmNodePrioritiesWarningValidator,
+            queue_name=self.name,
+            compute_resources=self.compute_resources,
         )
         if any(isinstance(compute_resource, SlurmComputeResource) for compute_resource in self.compute_resources):
             self._register_validator(
@@ -3095,9 +3217,16 @@ class SchedulerPluginClusterConfig(CommonSchedulerClusterConfig):
 class SlurmClusterConfig(CommonSchedulerClusterConfig):
     """Represent the full Slurm Cluster configuration."""
 
-    def __init__(self, cluster_name: str, scheduling: SlurmScheduling, **kwargs):
+    def __init__(
+        self,
+        cluster_name: str,
+        scheduling: SlurmScheduling,
+        login_nodes: LoginNodes = None,
+        **kwargs,
+    ):
         super().__init__(cluster_name, **kwargs)
         self.scheduling = scheduling
+        self.login_nodes = login_nodes
         self.__image_dict = None
         # Cache capacity reservations information together to reduce number of boto3 calls.
         # Since this cache is only used for validation, if AWSClientError happens
@@ -3148,6 +3277,7 @@ class SlurmClusterConfig(CommonSchedulerClusterConfig):
 
         instance_types_data = self.get_instance_types_data()
         self._register_validator(MultiNetworkInterfacesInstancesValidator, queues=self.scheduling.queues)
+
         for queue in self.scheduling.queues:
             for compute_resource in queue.compute_resources:
                 if self.scheduling.settings.enable_memory_based_scheduling:
@@ -3188,7 +3318,7 @@ class SlurmClusterConfig(CommonSchedulerClusterConfig):
                         InstancesEFAValidator,
                         InstancesNetworkingValidator,
                         InstancesAllocationStrategyValidator,
-                        InstancesMemorySchedulingValidator,
+                        InstancesMemorySchedulingWarningValidator,
                     ]
                     for validator in flexible_instance_types_validators:
                         self._register_validator(validator, **validator_args)