+++ CHANGELOG.md
@@ -5,8 +5,12 @@ CHANGELOG
 ------
 
 **ENHANCEMENTS**
+- Allow configuration of static and dynamic node priorities in Slurm compute resources via the ParallelCluster configuration YAML file.
+- Add support for Ubuntu 22.
+- Allow memory-based scheduling when multiple instance types are specified for a Slurm Compute Resource.
 
 **CHANGES**
+- Assign Slurm dynamic nodes a priority (weight) of 1000 by default. This allows Slurm to prioritize idle static nodes over idle dynamic ones.
 
 **BUG FIXES**
 
@@ -62,6 +66,7 @@ CHANGELOG
 - Upgrade image used by CodeBuild environment when building container images for AWS Batch clusters, from
   `aws/codebuild/amazonlinux2-x86_64-standard:3.0` to `aws/codebuild/amazonlinux2-x86_64-standard:4.0` and from
   `aws/codebuild/amazonlinux2-aarch64-standard:1.0` to `aws/codebuild/amazonlinux2-aarch64-standard:2.0`.
+- Avoid to reset FSx and EFS shared folders permissions when mounting them in the compute nodes.
 
 **BUG FIXES**
 - Fix EFS, FSx network security groups validators to avoid reporting false errors.